+++ R/effect.R
@@ -84,15 +84,16 @@ add_mappers <- function(...) {
 #' \itemize{
 #' \item{\eqn{c} }{is the *intercept*}
 #' \item{\eqn{x }}{is a *covariate*}
-#' \item{\eqn{\beta} }{is a *random variable* associated with \eqn{x} and}
-#' \item{\eqn{\psi = \beta * x }}{ is called the *random effect* of \eqn{x}}
+#' \item{\eqn{\beta} }{is a *latent variable* associated with \eqn{x} and}
+#' \item{\eqn{\psi = \beta * x }}{ is called the *effect* of \eqn{x}}
 #' }
 #'
 #' A problem that arises when using this kind of R formula is that it does not
 #' clearly reflect the mathematical
 #' formula. For instance, when providing the formula to inla, the resulting
 #' object will refer to the random
-#' effect \eqn{\psi = \beta * x } as `x`. Hence, it is not clear if `x` refers to the covariate
+#' effect \eqn{\psi = \beta * x } as `x`.
+#' Hence, it is not clear when `x` refers to the covariate
 #' or the effect of the covariate.
 #'
 #' @section Naming random effects:
@@ -174,14 +175,16 @@ component <- function(...) {
 #' extracts the mesh object to use as the mapper, and auto-generates mappers
 #' for indexed models. (Default: NULL, for auto-determination)
 #' @param main_layer,main_selector
-#' The `_layer` is a numeric index or character name of which
-#' layer/variable to extract from a covariate data object given in `main`
-#' (Default: The effect component name, if it exists i the covariate object,
-#' otherwise the first column of the covariate data frame)
+#' The `_layer` input should evaluate to a numeric index or character name or
+#' vector of which
+#' layer/variable to extract from a covariate data object given in `main`.
+#' (Default: NULL if `_selector` is given. Otherwise the effect component name,
+#'  if it exists in the covariate object, and otherwise the first column of
+#'  the covariate data frame)
 #'
-#' The `_selector` is character name of a variable whose contents determines which layer to
-#' extract from a covariate for each data point. Overrides the `layer`.
-#' (Default: NULL)
+#' The `_selector` value should be a character name of a variable
+#' whose contents determines which layer to extract from a covariate for each
+#' data point. (Default: NULL)
 #' @param n The number of latent variables in the model. Should be auto-detected
 #' for most or all models (Default: NULL, for auto-detection).
 #' An error is given if it can't figure it out by itself.
@@ -311,11 +314,6 @@ component.character <- function(object,
     group_model <- "exchangeable"
   }
 
-  if (is.null(main_layer)) {
-    main_layer <- label
-  }
-
-
   if ("map" %in% names(sys.call())) {
     #    if (!is.null(substitute(map))) {
     if (is.null(substitute(main))) {
@@ -367,7 +365,7 @@ component.character <- function(object,
       input = bru_input(
         substitute(main),
         label = label,
-        layer = main_layer,
+        layer = substitute(main_layer),
         selector = main_selector
       ),
       mapper = mapper,
@@ -380,7 +378,7 @@ component.character <- function(object,
       input = bru_input(
         substitute(group),
         label = paste0(label, ".group"),
-        layer = group_layer,
+        layer = substitute(group_layer),
         selector = group_selector
       ),
       mapper = group_mapper,
@@ -391,7 +389,7 @@ component.character <- function(object,
       input = bru_input(
         substitute(replicate),
         label = paste0(label, ".repl"),
-        layer = replicate_layer,
+        layer = substitute(replicate_layer),
         selector = replicate_selector
       ),
       mapper = replicate_mapper,
@@ -405,7 +403,7 @@ component.character <- function(object,
         bru_input(
           substitute(weights),
           label = paste0(label, ".weights"),
-          layer = weights_layer,
+          layer = substitute(weights_layer),
           selector = weights_selector
         )
       },
@@ -456,12 +454,15 @@ component.character <- function(object,
     component$replicate$mapper <- bru_mapper_index(1L)
     # Add scalable multi-mapper
     component[["mapper"]] <-
-      bru_mapper_scale(
-        bru_mapper_multi(list(
-          main = component$main$mapper,
-          group = component$group$mapper,
-          replicate = component$replicate$mapper
-        ))
+      bru_mapper_pipe(
+        list(
+          mapper = bru_mapper_multi(list(
+            main = component$main$mapper,
+            group = component$group$mapper,
+            replicate = component$replicate$mapper
+          )),
+          scale = bru_mapper_scale()
+        )
       )
   } else {
     if (!is.null(copy)) {
@@ -701,9 +702,15 @@ component_list.list <- function(object,
 
 #' @export
 #' @details * `c.component_list`: The `...` arguments should be `component_list`
-#' objects. The environment from the first argument will be applied to the resulting list.
+#' objects. The environment from the first argument will be applied to the
+#' resulting `component_list`.
 #' @rdname component_list
 `c.component_list` <- function(...) {
+  stopifnot(all(vapply(
+    list(...),
+    function(x) inherits(x, "component_list"),
+    TRUE
+  )))
   env <- environment(list(...)[[1]])
   object <- NextMethod()
   class(object) <- c("component_list", "list")
@@ -712,7 +719,25 @@ component_list.list <- function(object,
 }
 
 #' @export
-#' @param x `component_list` object from which to extract element(s)
+#' @details * `c.component`: The `...` arguments should be `component`
+#' objects. The environment from the first argument will be applied to the
+#' resulting ``component_list`.
+#' @rdname component_list
+`c.component` <- function(...) {
+  stopifnot(all(vapply(
+    list(...),
+    function(x) inherits(x, "component"),
+    TRUE
+  )))
+  env <- environment(list(...)[[1]])
+  object <- list(...)
+  class(object) <- c("component_list", "list")
+  environment(object) <- env
+  object
+}
+
+#' @export
+#' @param x `component_list` object from which to extract a sub-list
 #' @param i indices specifying elements to extract
 #' @rdname component_list
 `[.component_list` <- function(x, i) {
@@ -758,19 +783,22 @@ add_mappers.component <- function(component, lhoods, ...) {
     )
   lh <- lhoods[keep_lh]
 
-  component$main <- add_mapper(component$main,
+  component$main <- add_mapper(
+    component$main,
     label = component$label,
     lhoods = lh,
     env = component$env,
     require_indexed = FALSE
   )
-  component$group <- add_mapper(component$group,
+  component$group <- add_mapper(
+    component$group,
     label = component$label,
     lhoods = lh,
     env = component$env,
     require_indexed = TRUE
   )
-  component$replicate <- add_mapper(component$replicate,
+  component$replicate <- add_mapper(
+    component$replicate,
     label = component$label,
     lhoods = lh,
     env = component$env,
@@ -778,12 +806,15 @@ add_mappers.component <- function(component, lhoods, ...) {
   )
   # Add scalable multi-mapper
   component[["mapper"]] <-
-    bru_mapper_scale(
-      bru_mapper_multi(list(
-        main = component$main$mapper,
-        group = component$group$mapper,
-        replicate = component$replicate$mapper
-      ))
+    bru_mapper_pipe(
+      list(
+        mapper = bru_mapper_multi(list(
+          main = component$main$mapper,
+          group = component$group$mapper,
+          replicate = component$replicate$mapper
+        )),
+        scale = bru_mapper_scale()
+      )
     )
 
   fcall <- component$fcall
@@ -1130,7 +1161,8 @@ make_submapper <- function(subcomp_n,
                            label,
                            subcomp_type,
                            subcomp_factor_mapping,
-                           require_indexed) {
+                           require_indexed,
+                           allow_interpolation = TRUE) {
   if (!is.null(subcomp_n)) {
     if (!is.null(subcomp_values)) {
       warning(
@@ -1157,15 +1189,27 @@ make_submapper <- function(subcomp_n,
   } else {
     values <- sort(unique(values), na.last = NA)
     if (length(values) > 1) {
-      mapper <-
-        bru_mapper(
-          INLA::inla.mesh.1d(values),
-          indexed = require_indexed
-        )
+      if (allow_interpolation) {
+        mapper <-
+          bru_mapper(
+            INLA::inla.mesh.1d(values),
+            indexed = require_indexed
+          )
+      } else {
+        mapper <- bru_mapper_index(n = ceiling(max(values)))
+      }
     } else if (all(values == 1)) {
-      mapper <- bru_mapper_linear()
+      if (require_indexed) {
+        mapper <- bru_mapper_index(n = 1)
+      } else {
+        mapper <- bru_mapper_linear()
+      }
     } else {
-      mapper <- bru_mapper_linear()
+      if (require_indexed) {
+        mapper <- bru_mapper_factor(values, factor_mapping = "full")
+      } else {
+        mapper <- bru_mapper_linear()
+      }
     }
   }
 
@@ -1314,6 +1358,8 @@ make_mapper <- function(subcomp,
     subcomp[["mapper"]] <- bru_mapper_matrix(labels)
   } else if (subcomp[["model"]] %in% c("bym", "bym2")) {
     # No mapper; construct based on input values
+    # Default mapper will be integer or factor indexed, not
+    # interpolated
     mappers <- list(
       make_submapper(
         subcomp_n = subcomp[["n"]],
@@ -1322,7 +1368,8 @@ make_mapper <- function(subcomp,
         label = paste0(subcomp[["label"]], " part 1"),
         subcomp_type = subcomp[["type"]],
         subcomp_factor_mapping = subcomp[["factor_mapping"]],
-        require_indexed = require_indexed
+        require_indexed = require_indexed,
+        allow_interpolation = FALSE
       )
     )
     mappers[[2]] <- mappers[[1]]
@@ -1331,6 +1378,7 @@ make_mapper <- function(subcomp,
       bru_mapper_collect(mappers, hidden = TRUE)
   } else {
     # No mapper; construct based on input values
+    # Interpolation on by default
     subcomp[["mapper"]] <-
       make_submapper(
         subcomp_n = subcomp[["n"]],
@@ -1339,7 +1387,8 @@ make_mapper <- function(subcomp,
         label = subcomp[["label"]],
         subcomp_type = subcomp[["type"]],
         subcomp_factor_mapping = subcomp[["factor_mapping"]],
-        require_indexed = require_indexed
+        require_indexed = require_indexed,
+        allow_interpolation = TRUE
       )
   }
   subcomp
@@ -1432,6 +1481,9 @@ code.components <- function(components, add = "") {
   codes
 }
 
+
+
+
 # OPERATORS ----
 
 
@@ -1442,38 +1494,64 @@ code.components <- function(components, add = "") {
 #' @keywords internal
 #' @param object A `component` or `component_list`.
 #' @param ... ignored.
+#' @param depth The depth of which to expand the component mapper.
+#' Default `Inf`, to traverse the entire mapper tree.
 #' @author Fabian E. Bachl \email{bachlfab@@gmail.com}
+#' @author Finn Lindgren \email{finn.lindgren@@gmail.com}
 #'
 
-summary.component <- function(object, ...) {
+summary.component <- function(object, ..., depth = Inf) {
   result <- list(
     "Label" = object$label,
-    "  Main" = sprintf(
-      "input '%s',\ttype '%s'",
-      deparse(object[["main"]][["input"]][["input"]]),
-      object[["main"]][["type"]]
+    "Main" = sprintf(
+      "type '%s', input '%s'",
+      object[["main"]][["type"]],
+      deparse(object[["main"]][["input"]][["input"]])
     ),
-    "  Group" =
+    "Group" =
       if (!is.null(object[["group"]][["input"]][["input"]])) {
         sprintf(
-          "input '%s',\ttype '%s'",
-          deparse(object[["group"]][["input"]][["input"]]),
-          object[["group"]][["type"]]
+          "type '%s', input '%s'",
+          object[["group"]][["type"]],
+          deparse(object[["group"]][["input"]][["input"]])
         )
       } else {
         NULL
       },
-    "  Replicate" =
+    "Replicate" =
       if (!is.null(object[["replicate"]][["input"]][["input"]])) {
         sprintf(
-          "input '%s',\ttype '%s'",
-          deparse(object[["replicate"]][["input"]][["input"]]),
-          object[["replicate"]][["type"]]
+          "type '%s', input '%s'",
+          object[["replicate"]][["type"]],
+          deparse(object[["replicate"]][["input"]][["input"]])
         )
       } else {
         NULL
       },
-    "  INLA formula" = as.character(object$inla.formula)
+    "Mapper" =
+      if (is.null(object[["mapper"]])) {
+        "Not yet initialised"
+      } else {
+        strwrap(
+          summary(object[["mapper"]],
+            prefix = "    ",
+            initial = "",
+            depth = depth
+          ),
+          prefix = "    ",
+          initial = ""
+        )
+      },
+    "INLA formula" =
+      paste0(
+        "\n",
+        strwrap(
+          paste(as.character(object$inla.formula), collapse = " "),
+          width = 0.9 * getOption("width"),
+          indent = 4,
+          exdent = 6
+        )
+      )
   )
   if (!is.null(object[["copy"]])) {
     result[["Copy"]] <- sprintf("Copy of component '%s'", object[["copy"]])
@@ -1505,7 +1583,13 @@ summary.component_list <- function(object, ...) {
 print.summary_component <- function(x, ...) {
   for (name in names(x)) {
     if (!is.null(x[[name]])) {
-      cat(name, ":", "\t", x[[name]], "\n", sep = "")
+      if (name %in% "Label") {
+        cat("Label:", "\t", x[[name]], "\n", sep = "")
+      } else if (name %in% "Mapper") {
+        cat("  ", "Map: ", x[[name]], "\n", sep = "")
+      } else {
+        cat("  ", name, ":", "\t", x[[name]], "\n", sep = "")
+      }
     }
   }
   invisible(x)
@@ -1527,10 +1611,11 @@ print.summary_component_list <- function(x, ...) {
 #' @export
 #' @keywords internal
 #' @param component A component.
-#' @param data A `data.frame` or `Spatial*` object of covariates and/or point locations.
 #' @param input Component inputs, from `input_eval()`
-#' @param ... Unused.
-#' @return A `bru_mapper_taylor` object.
+#' @param linearisation evaluation state
+#' @param ... Optional parameters passed on to `ibm_eval`
+#' and `ibm_jacobian.
+#' @return A `bru_mapper_taylor` or `comp_simple_list` object.
 #' @author Finn Lindgren \email{finn.lindgren@@gmail.com}
 #' @rdname comp_lin_eval
 
@@ -1549,14 +1634,21 @@ comp_lin_eval.component <- function(component,
 
 comp_lin_eval.component_list <- function(components, input, state, ...) {
   # Note: Make sure the list element names carry over!
-  lapply(components,
-         function(x) {
-           label <- x[["label"]]
-           comp_lin_eval(x,
-                         input = input[[label]],
-                         state = state[[label]],
-                         ...)
-         })
+  mappers <-
+    lapply(
+      components,
+      function(x) {
+        label <- x[["label"]]
+        comp_lin_eval(x,
+          input = input[[label]],
+          state = state[[label]],
+          ...
+        )
+      }
+    )
+
+  class(mappers) <- c("comp_simple_list", class(mappers))
+  mappers
 }
 
 #' @section Simple covariates and the map parameter:
@@ -1630,17 +1722,17 @@ comp_lin_eval.component_list <- function(components, input, state, ...) {
 #' If `NULL`, return the component's map.
 #' @param ... Unused.
 #' @return An list of mapper input values, formatted for the full component mapper
-#' (of type `bru_mapper_scale`)
+#' (of type `bru_mapper_pipe`)
 #' @author Fabian E. Bachl \email{bachlfab@@gmail.com}, Finn Lindgren \email{finn.lindgren@@gmail.com}
 #' @rdname input_eval
 
 input_eval.component <- function(component,
                                  data,
                                  ...) {
-  stopifnot(inherits(component[["mapper"]], "bru_mapper_scale"))
+  stopifnot(inherits(component[["mapper"]], "bru_mapper_pipe"))
 
   # The names should be a subset of main, group, replicate
-  part_names <- names(component[["mapper"]][["mapper"]])
+  part_names <- ibm_names(component[["mapper"]][["mappers"]][[1]])
   mapper_val <- list()
   for (part in part_names) {
     mapper_val[[part]] <-
@@ -1681,10 +1773,37 @@ input_eval.component_list <-
 
 
 
+input_eval_layer <- function(layer, selector = NULL, envir, enclos,
+                             label,
+                             e_input) {
+  input_layer <- tryCatch(
+    eval(layer, envir = envir, enclos = enclos),
+    error = function(e) {
+      e
+    }
+  )
+  if (inherits(input_layer, "error")) {
+    stop(paste0(
+      "Failed to evaluate 'layer' input '",
+      deparse(layer),
+      "' for '",
+      paste0(label, ":layer"),
+      "'."
+    ))
+  }
+  if (is.null(input_layer) && is.null(selector)) {
+    if (label %in% names(e_input)) {
+      input_layer <- label
+    }
+  }
+  input_layer
+}
+
+
 #' @export
 #' @rdname input_eval
 
-input_eval.bru_input <- function(input, data, env = NULL, label = NULL,
+input_eval.bru_input <- function(input, data, env = NULL,
                                  null.on.fail = FALSE, ...) {
 
   # Evaluate the map with the data in an environment
@@ -1707,7 +1826,7 @@ input_eval.bru_input <- function(input, data, env = NULL, label = NULL,
   }
   assign(".data.", data, envir = envir)
 
-  emap <- tryCatch(eval(input$input, envir = envir, enclos = enclos),
+  e_input <- tryCatch(eval(input$input, envir = envir, enclos = enclos),
     error = function(e) {
     }
   )
@@ -1728,15 +1847,15 @@ input_eval.bru_input <- function(input, data, env = NULL, label = NULL,
   # ## A matrix.
   #  n <- nrow(as.data.frame(data))
 
-  if (is.null(emap)) {
+  if (is.null(e_input)) {
     if (null.on.fail) {
       return(NULL)
     }
     #    val <- rep(1, n)
     val <- 1
-  } else if (is.function(emap)) {
+  } else if (is.function(e_input)) {
     # Allow but detect failures:
-    val <- tryCatch(emap(data),
+    val <- tryCatch(e_input(data),
       error = function(e) {
       }
     )
@@ -1766,9 +1885,9 @@ input_eval.bru_input <- function(input, data, env = NULL, label = NULL,
         }
       )
     }
-  } else if (inherits(emap, "formula")) {
+  } else if (inherits(e_input, "formula")) {
     # Allow but detect failures:
-    val <- tryCatch(MatrixModels::model.Matrix(emap, data = data, sparse = TRUE),
+    val <- tryCatch(MatrixModels::model.Matrix(e_input, data = data, sparse = TRUE),
       error = function(e) {
       }
     )
@@ -1778,28 +1897,41 @@ input_eval.bru_input <- function(input, data, env = NULL, label = NULL,
       # that are likely to happen for multilikelihood models; A component only
       # needs to be evaluable for at least one of the likelihoods.
     }
-  } else if (inherits(emap, "SpatialGridDataFrame") ||
-    inherits(emap, "SpatialPixelsDataFrame")) {
-    if (is.null(input[["selector"]])) {
-      layer <-
-        if (is.null(input[["layer"]])) {
-          1
-        } else {
-          input[["layer"]]
-        }
-    }
-    val <- eval_SpatialDF(
-      emap,
+  } else if (inherits(
+    e_input,
+    c(
+      "SpatialGridDataFrame",
+      "SpatialPixelsDataFrame",
+      "SpatRaster"
+    )
+  )) {
+    input_layer <-
+      input_eval_layer(
+        layer = input[["layer"]],
+        selector = input[["selector"]],
+        envir = envir,
+        enclos = enclos,
+        label = input[["label"]],
+        e_input = e_input
+      )
+    layer <- extract_layer(
+      data,
+      input_layer,
+      input[["selector"]]
+    )
+    check_layer(e_input, data, layer)
+    val <- eval_spatial(
+      e_input,
       data,
       layer = layer,
-      selector = input[["selector"]]
+      selector = NULL
     )
     #  } else if ((input$label %in% c("offset", "const")) &&
-    #    is.numeric(emap) &&
-    #    (length(emap) == 1)) {
-    #    val <- rep(emap, n)
+    #    is.numeric(e_input) &&
+    #    (length(e_input) == 1)) {
+    #    val <- rep(e_input, n)
   } else {
-    val <- emap
+    val <- e_input
   }
 
   # ## Need to allow different sizes; K %*% effect needs to have same length as response, but the input and effect effect itself doesn't.
@@ -1812,14 +1944,18 @@ input_eval.bru_input <- function(input, data, env = NULL, label = NULL,
   # to fix that by filling in nearest neighbour values.
   # # TODO: Check how to deal with this fully in the case of multilikelihood models
   # Answer: should respect the lhood "include/exclude" info for the component list
-  if ((inherits(emap, "SpatialGridDataFrame") ||
-    inherits(emap, "SpatialPixelsDataFrame")) &&
+  if ((inherits(e_input, c(
+    "SpatialGridDataFrame",
+    "SpatialPixelsDataFrame",
+    "SpatRaster"
+  ))) &&
     any(is.na(as.data.frame(val)))) {
     warning(
       paste0(
         "Model input '",
         deparse(input$input),
-        "' for '", label, "' returned some NA values.\n",
+        "' for '", input$label,
+        "' returned some NA values.\n",
         "Attempting to fill in spatially by nearest available value.\n",
         "To avoid this basic covariate imputation, supply complete data."
       ),
@@ -1827,12 +1963,17 @@ input_eval.bru_input <- function(input, data, env = NULL, label = NULL,
     )
 
     val <- bru_fill_missing(
-      data = emap, where = data, values = val,
-      layer = layer, selector = input[["selector"]]
+      data = e_input, where = data, values = val,
+      layer = layer, selector = NULL
     )
   }
 
   # Check for NA values.
+  # TODO: update 2022-11-04:
+  # iid models now by default are given _index mappers, which treat NAs
+  # as zero.  An option to turn on a warning could be useful for checking
+  # models that aren't supposed to have NA inputs.
+  #
   #  if (any(is.na(unlist(as.vector(val))))) {
   #    # TODO: remove this check and make sure NAs are handled properly elsewhere,
   #    # if possible. Problem: treating NA as "no effect" can have negative side