+++ tests/testthat/test_predict.R
@@ -56,8 +56,8 @@ test_that("bru: factor component", {
     seed = 12345L
   )
 
-  # The statistics include mean, standard deviation, the 2.5% quantile, the median,
-  # the 97.5% quantile
+  # The default statistics include mean, standard deviation,
+  # the 2.5% quantile, the median, the 97.5% quantile
   expect_equal(is.data.frame(xpost), TRUE)
   expect_equal(nrow(xpost), 1)
 
@@ -103,7 +103,7 @@ test_that("bru: factor component", {
     xpost4[4, , drop = FALSE],
     xpost4[5, , drop = FALSE]
   )
-  # The columns should all be different
+  # The columns should all different values
   expect_equal(
     sum(xpost4[, 1] == xpost4[, 2]),
     0