+++ NEWS.md
@@ -1,6 +1,49 @@
 # inlabru (development version)
 
 ## Features
+
+* Allow `NA` input for default 1D mappers to generate effect zero, like
+  in `inla()`.
+
+* Further `bru_mapper` method updates;
+
+  * Deprecated `ibm_amatrix()` and `names()`
+  methods, replaced by `ibm_jacobian()` and `ibm_names()`.
+  
+  * Introduced `bru_mapper_pipe()`, used to link mappers in sequence.
+  
+  * Introduced `bru_mapper_aggregate()` and `bru_mapper_logsumexp()`,
+    used for blockwise weighted sums and log-sum-exp mappings,
+    `output[k] = sum(weights[block==k]*state[block==k])))` and
+    `output[k] = log(sum(weights[block==k]*exp(state[block==k])))`,
+    with optional weight normalisation within each block.  Allows
+    providing the weights as log-weights, and uses block-wise shifts to
+    avoid potential overflow.
+  
+  * `summary` methods for `bru_mapper` objects
+  
+  * Removed `methods` argument from `bru_mapper_define()`.  Implementations
+    should register S3 methods instead.
+
+## Bug fixes
+
+* Remove unused `spatstat.core` dependency. Fixes #165
+
+* Fixed issue with plain mapper evaluation in the `ibm_eval.default`
+  and `ibm_eval.bru_mapper_collect` methods, where they would return zeros
+  instead of the intended values.
+  The main component evaluation and estimation code was not directly affected
+  as that is based on the `bru_mapper_multi` class methods that rely on the
+  Jacobians instead.  The bug would therefore mainly have impacted the future,
+  not yet supported nonlinear mapper extensions.
+  
+* Fix for eval_spatial.SpatRaster; Work around inconsistent logic in
+  `terra::extract(..., layer)` when `length(layer)==1` or `nrow(where)==1`.
+  Fixes #169
+
+# inlabru 2.6.0
+
+## Features
   
 * Add `bru_get_mapper` generic, and associated methods for `inla.spde` and
   `inla.rgeneric` objects. This allows `inlabru` to automatically extract
@@ -14,6 +57,14 @@
 * The `<component>_eval()` methods available in predictor expressions
   now handle optional scaling weights, like in ordinary component effect
   evaluation.
+
+* Add `terra` support for covariate inputs
+
+* The component `*_layer` arguments are now evaluated in the data context,
+  to allow dynamic layer selection for spatial raster covariates.  A new
+  generic `eval_spatial()` provides support for grid/pixel based
+  `Spatial*DataFrame` evaluation, and `SpatRaster`. Expanded support
+  is in progress.
 
 * New vignettes on the `bru_mapper` system, `component` definitions,
   and `prediction_scores`
@@ -34,15 +85,16 @@
   * New mapper class `bru_mapper_const`, which replaces `bru_mapper_offset`.
   `bru_mapper_offset` is now deprecated and will produce warnings.
   
-  
 ## Bug fixes
 
 * Enable both datum/ensemble container for ellipsoid information, to support
   `epsg:4326`. Fixes #154
   
-* Make duplicated component names an error instead of a warning. Relates to #155
+* Make duplicated component names an error instead of a warning.
+  Relates to #155
 
-* Fix Tsparse assumptions in `row_kron`. Fixes #162
+* Fix `Tsparse` assumptions in `row_kron` to prepare for Matrix `1.5-2`.
+  Fixes #162
   
 # inlabru 2.5.3